+++ src/main/java/com/c_comachi/inused/domain/chat/controller/ChatController.java
@@ -1,48 +1,32 @@
 package com.c_comachi.inused.domain.chat.controller;
 
 
-import com.c_comachi.inused.domain.chat.moel.ChatMessage;
-import com.c_comachi.inused.domain.chat.repository.ChatRoomRepository;
+import com.c_comachi.inused.domain.chat.dto.request.ChatMessageRequestDto;
+import com.c_comachi.inused.domain.chat.entity.ChatMessage;
+import com.c_comachi.inused.domain.chat.service.ChatMessageService;
 import com.c_comachi.inused.domain.users.entity.UserEntity;
-import com.c_comachi.inused.domain.users.jwt.TokenProvider;
 import com.c_comachi.inused.domain.users.repository.UserRepository;
-import com.c_comachi.inused.global.service.RedisPublisher;
-import java.security.Principal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 
 @RequiredArgsConstructor
 @Controller
 public class ChatController {
 
-    private final RedisTemplate<String, Object> redisTemplate;
-    private final ChannelTopic channelTopic;
-    private final UserRepository userRepository;
+    private final ChatMessageService chatMessageService;
 
     /**
      * websocket "/pub/chat/message"로 들어오는 메시징을 처리한다.
      */
     @MessageMapping("/chat/message")
-    public void message(@Header(value = "nickname") String nickname, ChatMessage message) {
-        UserEntity user = userRepository.findByEmail(nickname).get();
+    public void message(@Header(value = "email") String email, ChatMessageRequestDto chatMessageRequestDto) {
 
-       String realName = user.getNickname();
-        // 로그인 회원 정보로 대화명 설정
-        message.setSender(realName);
-        // 채팅방 입장시에는 대화명과 메시지를 자동으로 세팅한다.
-        if (ChatMessage.MessageType.ENTER.equals(message.getType())) {
-            message.setSender("[알림]");
-            message.setMessage(realName + "님이 입장하셨습니다.");
-        }
-
+        chatMessageService.sendMessage(chatMessageRequestDto, email);
         // Websocket에 발행된 메시지를 redis로 발행(publish)
-        redisTemplate.convertAndSend(channelTopic.getTopic(), message);
     }
 }
