+++ src/main/java/com/c_comachi/inused/global/config/WebSockConfig.java
@@ -18,13 +18,14 @@ public class WebSockConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void configureMessageBroker(MessageBrokerRegistry config) {
-        config.enableSimpleBroker("/sub");
-        config.setApplicationDestinationPrefixes("/pub");
+        config.enableSimpleBroker("/sub"); // 해당 주소를 구독하는 클라어인트에게 메시지를 보낸다.
+        config.setApplicationDestinationPrefixes("/pub"); // /pub 으로 시작하는 메세지만 Broker에서 처리
     }
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/ws-stomp").setAllowedOrigins("http://localhost:8080")
+        registry.addEndpoint("/ws-stomp") // ex ) ws://localhost:8080/ws-stomp
+                .setAllowedOrigins("http://localhost:8080") // 접근할 수 있는 권한
                 .withSockJS();
     }
