+++ src/main/java/com/c_comachi/inused/global/service/RedisService.java
@@ -1,12 +1,10 @@
 package com.c_comachi.inused.global.service;
 
-import com.c_comachi.inused.domain.chat.moel.ChatMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
